requirements.txt
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
python-dotenv
psycopg2
Pillow